finished need to add lock for store
